Add ROS2 Dashing bundle support
This adds support for ROS2 Dashing in the `colcon bundle` command.
A new integration test workspace has been added for ROS2 since it
builds using a different ecosystem than ROS1.

+++ run_integration_test.sh
@@ -2,6 +2,12 @@
 
 set -ex
 
+if [[ ${ROS_DISTRO} = "dashing" ]]
+then
+	./run_ros2_integration_test.sh
+	exit 0
+fi
+
 # Clean up
 rm -rf integration/v1_bundle
 rm -rf integration/v2_bundle
